perf(annotation): equality-gate frame track selection; extract useFrameDerivedTracks
