+++ app/packages/video-annotation/src/components/FrameLabels.tsx
@@ -1,11 +1,6 @@
-import {
-  useActiveSampleId,
-  useAnnotationEngine,
-  useEngineSelector,
-} from "@fiftyone/annotation";
 import { getLabelColorFromContext } from "@fiftyone/lighter";
 import type { ModalSample } from "@fiftyone/state";
-import type { LabelData, Stage } from "@fiftyone/utilities";
+import type { Stage } from "@fiftyone/utilities";
 import React, { useCallback, useEffect, useMemo, useRef } from "react";
 import {
   useActiveDetectionField,
@@ -36,14 +31,15 @@ import {
   usePublishFrameLabelsStream,
 } from "../streams/frameLabelsStream";
 import {
-  buildTracksFromIndex,
   objectTrackClassOf,
   objectTrackPathOf,
   parseSubTrackId,
-  type FrameOverlay,
   type PerInstanceLabel,
 } from "../tracks/frameTracks";
-import { useVideoLabelsIndex } from "../hooks/useVideoLabelsIndex";
+import {
+  useFrameDerivedTracks,
+  type ObjectTrackColorResolver,
+} from "../tracks/useFrameDerivedTracks";
 import {
   useTrackExpansion,
   type TrackExpansion,
@@ -96,12 +92,6 @@ type TrackDecoration = BaseTrackDecoration & {
 
 /** Row height (px) for a dynamic-attribute sub-track — shorter than a parent. */
 const SUB_TRACK_ROW_HEIGHT = 22;
-
-/** Resolves the row color for a per-frame object track. */
-type ObjectTrackColorResolver = (
-  label: PerInstanceLabel,
-  path: string,
-) => string;
 
 /** Resolves the row color for a temporal-detection track. */
 type TemporalDetectionColorResolver = (
@@ -251,121 +241,6 @@ const FrameLabelsRegistration: React.FC<FrameLabelsRegistrationProps> = ({
 
   return <>{children}</>;
 };
-
-/**
- * Build the per-instance object tracks from the server distribution index
- * merged with the engine's edited-frame overlay — no whole-clip walk. Tracks
- * rebuild on each engine commit (cheap: only the dirty frames are read).
- * `resolved` flips true once the index settles, gating the pin bootstrap.
- */
-function useFrameDerivedTracks(
-  resolveColor: ObjectTrackColorResolver,
-  getDynamicAttributes: (path: string | null) => string[],
-): {
-  tracks: Track[];
-  resolved: boolean;
-} {
-  const stream = useFrameLabelsStream();
-  const engine = useAnnotationEngine();
-  const sampleId = useActiveSampleId();
-  const visible = useVisibleLabelSchemas();
-  const labelTypes = useFrameLabelFields();
-
-  // Fetch the index for every declared frame label field (stable per dataset/
-  // view) so visibility toggles filter client-side without re-fetching.
-  const allFields = useMemo(() => Object.keys(labelTypes), [labelTypes]);
-
-  // Each field's declared dynamic attributes, scoped per path so a field
-  // without dynamic attributes gets no sub-tracks.
-  const dynamicByPath = useMemo(() => {
-    const map: Record<string, string[]> = {};
-    for (const path of allFields) {
-      map[path] = getDynamicAttributes(path);
-    }
-    return map;
-  }, [allFields, getDynamicAttributes]);
-
-  // The fetch takes the union across fields — the backend returns attribute
-  // segments only for the fields that actually declare each attribute.
-  const allDynamicAttributes = useMemo(
-    () => Array.from(new Set(Object.values(dynamicByPath).flat())),
-    [dynamicByPath],
-  );
-
-  // Gate each frame field's tracks on the sidebar's visible set — deactivating
-  // a field in the schema manager hides its timeline rows, matching the canvas
-  // + sidebar.
-  const paths = useMemo(
-    () => allFields.filter((p) => visible.has(p)),
-    [allFields, visible],
-  );
-
-  const { indexByPath, loaded } = useVideoLabelsIndex(
-    stream,
-    allFields,
-    allDynamicAttributes,
-  );
-
-  // engine version is the rebuild signal.
-  const engineVersion = useEngineSelector(engine, () => engine.getVersion());
-
-  // No visible frame field, or the index hasn't settled: no rows. Tracks build
-  // per visible field from that field's index ⊕ its dirty-frame overlay, then
-  // concatenate — instance ids are unique across fields, so the rows just merge.
-  const tracks = useMemo(() => {
-    if (!stream || !sampleId || !loaded || paths.length === 0) {
-      return [];
-    }
-
-    return paths.flatMap((path) =>
-      buildTracksFromIndex({
-        path,
-        index: indexByPath[path] ?? [],
-        overlay: readEngineOverlay(engine, sampleId, path),
-        fps: stream.fps,
-        resolveColor,
-        dynamicAttributes: dynamicByPath[path] ?? [],
-      }),
-    );
-    // eslint-disable-next-line react-hooks/exhaustive-deps -- engineVersion is the invalidation signal
-  }, [
-    engine,
-    sampleId,
-    paths,
-    stream,
-    resolveColor,
-    indexByPath,
-    loaded,
-    dynamicByPath,
-    engineVersion,
-  ]);
-
-  return { tracks, resolved: loaded };
-}
-
-/**
- * The engine's materialized frames + their live labels, keyed by frame number —
- * the overlay that shadows the server index. Reads every loaded frame, not just
- * the dirty set: a successful autosave folds edits into the seed and clears the
- * dirty set, so a dirty-only overlay would revert the timeline to the stale
- * index after each save. The engine is authoritative for every frame it holds,
- * so overlaying all of them keeps the timeline correct post-save and composes
- * index (unloaded) ⊕ engine (loaded window) once the seed is windowed. Bounded
- * by the loaded window, which today is the whole clip (see `warmupAll`).
- */
-function readEngineOverlay(
-  engine: ReturnType<typeof useAnnotationEngine>,
-  sample: string,
-  path: string,
-): FrameOverlay {
-  const overlay: FrameOverlay = new Map<number, LabelData[]>();
-
-  for (const frame of engine.loadedFrames(sample)) {
-    overlay.set(frame, engine.listLabels({ sample, path, frame }));
-  }
-
-  return overlay;
-}
 
 /**
  * Derive TD tracks from the engine — the authoritative, reactive TD source.
